First part of major refactoring, working to supporting different types of serializers, and subschema's
