+++ Cargo.toml
@@ -17,16 +17,20 @@ default = []
 kafka_test = []
 
 [dependencies.avro-rs]
-version = ">= 0.6"
+version = ">= 0.10"
 
 [dependencies.byteorder]
-version = ">= 1.0"
+version = ">= 1.3"
 
 [dependencies.curl]
 version = ">= 0.4"
 
+[dependencies.protobuf]
+version = ">= 2.15"
+features = ["with-bytes"]
+
 [dependencies.failure]
-version = ">= 0.1.6"
+version = ">= 0.1.0"
 
 [dependencies.serde]
 version = "1.0"
@@ -39,10 +43,10 @@ version = ">=0.8.0, <2.0"
 version = ">= 2.1.1"
 
 [dev-dependencies]
-mockito = "0.23.3"
-rdkafka = { version = "0.23", features = ["cmake-build"] }
-futures = "^0.3.1"
-rand = "0.6.5"
+mockito = "0.26.0"
+rdkafka = { version = "0.23.1", features = ["cmake-build"] }
+futures = "^0.3.5"
+rand = "0.7.3"
 
 [badges.travis-ci]
 repository = "gklijs/schema_registry_converter"
