+++ src/lib.rs
@@ -22,1181 +22,39 @@
 //!
 //! [avro-rs]: https://crates.io/crates/avro-rs
 
-#[macro_use] extern crate failure;
+#[macro_use]
+extern crate failure;
 
+pub mod avro;
+mod raw;
 pub mod schema_registry;
-mod value_fixed_fix;
 
-use crate::value_fixed_fix::FixedFixer;
-use avro_rs::schema::Name;
-use avro_rs::to_value;
-use avro_rs::types::{Record, ToAvro, Value};
-use avro_rs::{from_avro_datum, to_avro_datum, Schema};
-use byteorder::{BigEndian, ByteOrder, ReadBytesExt};
-use schema_registry::SRCError;
-use schema_registry::{get_schema_by_id, get_schema_by_subject, get_subject, SubjectNameStrategy};
-use serde::ser::Serialize;
-use std::collections::hash_map::RandomState;
-use std::collections::HashMap;
-use std::io::Cursor;
+use crate::avro::{AvroDecoder, AvroEncoder};
 
-/// A decoder used to transform bytes to a Value object
-///
-/// The main purpose of having this struct is to be able to cache the schema's. Because the need to
-/// be retrieved over http from the schema registry, and we can already see from the bytes which
-/// schema we should use, this can save a lot of unnecessary calls.
-/// Errors are also stored to the cache, because they may not be recoverable. A function is
-/// available to remove the errors from the cache. To get the value avro_rs is used.
-///
-/// For both the key and the payload/key it's possible to use the schema registry, this struct supports
-/// both. But only using the SubjectNameStrategy::TopicNameStrategy it has to be made explicit
-/// whether it's actual used as key or value.
-///
-/// ```
-///  # use mockito::{mock, server_address};
-///  # use schema_registry_converter::Decoder;
-///  # use avro_rs::types::Value;
-///
-/// let _m = mock("GET", "/schemas/ids/1")
-///     .with_status(200)
-///     .with_header("content-type", "application/vnd.schemaregistry.v1+json")
-///     .with_body(r#"{"schema":"{\"type\":\"record\",\"name\":\"Heartbeat\",\"namespace\":\"nl.openweb.data\",\"fields\":[{\"name\":\"beat\",\"type\":\"long\"}]}"}"#)
-///     .create();
-///
-/// let mut decoder = Decoder::new(format!("http://{}", server_address()));
-/// let heartbeat = decoder.decode(Some(&[0,0,0,0,1,6]));
-///
-/// assert_eq!(heartbeat, Ok(Value::Record(vec!(("beat".to_string(), Value::Long(3))))))
-/// ```
 #[derive(Debug)]
 pub struct Decoder {
     schema_registry_url: String,
-    cache: &'static mut HashMap<u32, Result<Schema, SRCError>, RandomState>,
 }
 
 impl Decoder {
-    /// Creates a new decoder which will use the supplied url to fetch the schema's since the schema
-    /// needed is encoded in the binary, independent of the SubjectNameStrategy we don't need any
-    /// additional data. It's possible for recoverable errors to stay in the cash, when a result
-    /// comes back as an error you can use remove_errors_from_cache to clean the cache, keeping the
-    /// correctly fetched schema's
-    pub fn new(schema_registry_url: String) -> Decoder {
-        let new_cache = Box::new(HashMap::new());
-        Decoder {
-            schema_registry_url,
-            cache: Box::leak(new_cache),
-        }
+    pub fn new(schema_registry_url: String) -> AvroDecoder {
+        AvroDecoder::new(schema_registry_url)
     }
-    /// Remove al the errors from the cache, you might need to/want to run this when a recoverable
-    /// error is met. Errors are also cashed to prevent trying to get schema's that either don't
-    /// exist or can't be parsed.
-    ///
-    /// ```
-    ///  # use mockito::{mock, server_address};
-    ///  # use schema_registry_converter::Decoder;
-    ///  # use schema_registry_converter::schema_registry::SRCError;
-    ///  # use avro_rs::types::Value;
-    ///
-    /// let mut decoder = Decoder::new(format!("http://{}", server_address()));
-    /// let bytes = [0,0,0,0,2,6];
-    ///
-    /// let _m = mock("GET", "/schemas/ids/2")
-    ///     .with_status(404)
-    ///     .with_header("content-type", "application/vnd.schemaregistry.v1+json")
-    ///     .with_body(r#"{"error_code":40403,"message":"Schema not found"}"#)
-    ///     .create();
-    /// let heartbeat = decoder.decode(Some(&bytes));
-    /// assert_eq!(heartbeat, Err(SRCError::new("Did not get a 200 response code but 404 instead", None, false).into_cache()));
-    /// let _m = mock("GET", "/schemas/ids/2")
-    ///     .with_status(200)
-    ///     .with_header("content-type", "application/vnd.schemaregistry.v1+json")
-    ///     .with_body(r#"{"schema":"{\"type\":\"record\",\"name\":\"Heartbeat\",\"namespace\":\"nl.openweb.data\",\"fields\":[{\"name\":\"beat\",\"type\":\"long\"}]}"}"#)
-    ///     .create();
-    ///
-    /// let heartbeat = decoder.decode(Some(&bytes));
-    /// assert_eq!(heartbeat, Err(SRCError::new("Did not get a 200 response code but 404 instead", None, false).into_cache()));
-    ///
-    /// decoder.remove_errors_from_cache();
-    ///
-    /// let heartbeat = decoder.decode(Some(&bytes));
-    /// assert_eq!(heartbeat, Ok(Value::Record(vec!(("beat".to_string(), Value::Long(3))))))
-    /// ```
-    pub fn remove_errors_from_cache(&mut self) {
-        self.cache.retain(|_, v| v.is_ok());
-    }
-    /// Decodes bytes into a value.
-    /// The choice to use Option<&[u8]> as type us made so it plays nice with the BorrowedMessage
-    /// struct from rdkafka, for example if we have m: &'a BorrowedMessage and decoder: &'a mut
-    /// Decoder we can use decoder.decode(m.payload()) to decode the payload or
-    /// decoder.decode(m.key()) to get the decoded key.
-    ///
-    /// ```no_run
-    /// # use rdkafka::message::{Message, BorrowedMessage};
-    /// # use schema_registry_converter::Decoder;
-    /// # use avro_rs::types::Value;
-    /// fn get_value<'a>(
-    ///     msg: &'a BorrowedMessage,
-    ///     decoder: &'a mut Decoder,
-    /// ) -> Value{
-    ///     match decoder.decode(msg.payload()){
-    ///         Ok(v) => v,
-    ///         Err(e) => panic!("Error getting value: {}", e),
-    ///     }
-    /// }
-    /// ```
-    pub fn decode(&mut self, bytes: Option<&[u8]>) -> Result<Value, SRCError> {
-        match bytes {
-            None => Ok(Value::Null),
-            Some(p) if p.len() > 4 && p[0] == 0 => self.deserialize(p),
-            Some(p) => Ok(Value::Bytes(p.to_vec())),
-        }
-    }
-    /// The actual deserialization trying to get the id from the bytes to retrieve the schema, and
-    /// using a reader transforms the bytes to a value.
-    fn deserialize<'a>(&'a mut self, bytes: &'a [u8]) -> Result<Value, SRCError> {
-        let schema = self.get_schema(bytes);
-        let mut reader = Cursor::new(&bytes[5..]);
-        match schema {
-            Ok(v) => match from_avro_datum(&v, &mut reader, None) {
-                Ok(v) => Ok(v),
-                Err(e) => Err(SRCError::non_retryable_from_err(
-                    e,
-                    "Could not transform bytes using schema",
-                )),
-            },
-            Err(e) => Err(e.clone()),
-        }
-    }
-
-    /// Decodes bytes into a value.
-    /// Also gives back the name of  the schema used as a way to match with a struct to deserialize.
-    pub fn decode_with_name(&mut self, bytes: Option<&[u8]>) -> Result<(Name, Value), SRCError> {
-        match bytes {
-            None => Ok((Name::new("null"), Value::Null)),
-            Some(p) if p.len() > 4 && p[0] == 0 => self.deserialize_with_name(p),
-            Some(p) => Ok((Name::new("bytes"), Value::Bytes(p.to_vec()))),
-        }
-    }
-    /// The actual deserialization trying to get the id from the bytes to retrieve the schema, and
-    /// using a reader transforms the bytes to a value.
-    fn deserialize_with_name<'a>(&'a mut self, bytes: &'a [u8]) -> Result<(Name, Value), SRCError> {
-        let schema = self.get_schema(bytes);
-        let mut reader = Cursor::new(&bytes[5..]);
-        match schema {
-            Ok(v) => match from_avro_datum(&v, &mut reader, None) {
-                Ok(val) => Ok((get_name(v), val)),
-                Err(e) => Err(SRCError::new(
-                    "Could not transform bytes using schema",
-                    Some(&e.to_string()),
-                    false,
-                )),
-            },
-            Err(e) => Err(e.clone()),
-        }
-    }
-
-    fn get_schema(&mut self, bytes: &[u8]) -> &mut Result<Schema, SRCError> {
-        let mut buf = &bytes[1..5];
-        let id = buf.read_u32::<BigEndian>().unwrap();
-        let sr = &self.schema_registry_url;
-        self.cache
-            .entry(id)
-            .or_insert_with(|| match get_schema_by_id(id, sr) {
-                Ok(v) => Ok(v),
-                Err(e) => Err(e.into_cache()),
-            })
+    pub fn avro(schema_registry_url: String) -> AvroDecoder {
+        AvroDecoder::new(schema_registry_url)
     }
 }
 
-/// An encoder used to transform a Value object to bytes
-///
-/// The main purpose of having this struct is to be able to cache the schema's. Because the need to
-/// be retrieved over http from the schema registry, and we can already see from the bytes which
-/// schema we should use, this can save a lot of unnecessary calls.
-/// Errors are also stored to the cache, because they may not be recoverable. A function is
-/// available to remove the errors from the cache. To get the value avro_rs is used.
-///
-/// For both the key and the payload/key it's possible to use the schema registry, this struct supports
-/// both. But only using the SubjectNameStrategy::TopicNameStrategy it has to be made explicit
-/// whether it's actual used as key or value.
-///
-/// ```
-///  # use mockito::{mock, server_address};
-///  # use schema_registry_converter::Encoder;
-///  # use schema_registry_converter::schema_registry::SubjectNameStrategy;
-///  # use avro_rs::types::Value;
-///
-/// let _m = mock("GET", "/subjects/heartbeat-value/versions/latest")
-///     .with_status(200)
-///     .with_header("content-type", "application/vnd.schemaregistry.v1+json")
-///     .with_body(r#"{"subject":"heartbeat-value","version":1,"id":3,"schema":"{\"type\":\"record\",\"name\":\"Heartbeat\",\"namespace\":\"nl.openweb.data\",\"fields\":[{\"name\":\"beat\",\"type\":\"long\"}]}"}"#)
-///     .create();
-///
-/// let _m = mock("GET", "/subjects/heartbeat-key/versions/latest")
-///     .with_status(200)
-///     .with_header("content-type", "application/vnd.schemaregistry.v1+json")
-///     .with_body(r#"{"subject":"heartbeat-value","version":1,"id":4,"schema":"{\"type\":\"record\",\"name\":\"Name\",\"namespace\":\"nl.openweb.data\",\"fields\":[{\"name\":\"name\",\"type\":\"string\",\"avro.java.string\":\"String\"}]}"}"#)
-///     .create();
-///
-/// let mut encoder = Encoder::new(server_address().to_string());
-///
-/// let key_strategy = SubjectNameStrategy::TopicNameStrategy("heartbeat".into(), true);
-/// let bytes = encoder.encode(vec!(("name", Value::String("Some name".to_owned()))), &key_strategy);
-///
-/// assert_eq!(bytes, Ok(vec!(0, 0, 0, 0, 4, 18, 83, 111, 109, 101, 32, 110, 97, 109, 101)));
-///
-/// let value_strategy = SubjectNameStrategy::TopicNameStrategy("heartbeat".into(), false);
-/// let bytes = encoder.encode(vec!(("beat", Value::Long(3))), &value_strategy);
-///
-/// assert_eq!(bytes, Ok(vec!(0,0,0,0,3,6)))
-/// ```
 #[derive(Debug)]
 pub struct Encoder {
     schema_registry_url: String,
-    cache: &'static mut HashMap<String, Result<(Schema, u32), SRCError>, RandomState>,
 }
 
 impl Encoder {
-    /// Creates a new encoder which will use the supplied url to fetch the schema's. The schema's
-    /// need to be retrieved together with the id, in order for a consumer to decode the bytes.
-    /// For the encoding several strategies are available in the java client, all three of them are
-    /// supported. The schema's does have to be present in the schema registry already. This is
-    /// unlike the Java client with wich it's possible to update/upload the schema when it's not
-    /// present yet. While it may be added to this library, it's also not hard to do it separately.
-    /// New schema's can set by doing a post at /subjects/{subject}/versions.
-    pub fn new(schema_registry_url: String) -> Encoder {
-        let new_cache = Box::new(HashMap::new());
-        Encoder {
-            schema_registry_url,
-            cache: Box::leak(new_cache),
-        }
+    pub fn new(schema_registry_url: String) -> AvroEncoder {
+        AvroEncoder::new(schema_registry_url)
     }
-    /// Remove al the errors from the cache, you might need to/want to run this when a recoverable
-    /// error is met. Errors are also cashed to prevent trying to get schema's that either don't
-    /// exist or can't be parsed.
-    ///
-    /// ```
-    ///  # use mockito::{mock, server_address};
-    ///  # use schema_registry_converter::Encoder;
-    ///  # use schema_registry_converter::schema_registry::SubjectNameStrategy;
-    ///  # use schema_registry_converter::schema_registry::SRCError;
-    ///  # use avro_rs::types::Value;
-    ///
-    /// let mut encoder = Encoder::new(server_address().to_string());
-    /// let strategy = SubjectNameStrategy::RecordNameStrategy("nl.openweb.data.Heartbeat".into());
-    ///
-    /// let _m = mock("GET", "/subjects/nl.openweb.data.Heartbeat/versions/latest")
-    ///     .with_status(404)
-    ///     .with_header("content-type", "application/vnd.schemaregistry.v1+json")
-    ///     .with_body(r#"{"error_code":40403,"message":"Schema not found"}"#)
-    ///     .create();
-    ///
-    /// let bytes = encoder.encode(vec!(("beat", Value::Long(3))), &strategy);
-    /// assert_eq!(bytes, Err(SRCError::new("Did not get a 200 response code but 404 instead", None, false).into_cache()));
-    ///
-    /// let _m = mock("GET", "/subjects/nl.openweb.data.Heartbeat/versions/latest")
-    ///     .with_status(200)
-    ///     .with_header("content-type", "application/vnd.schemaregistry.v1+json")
-    ///     .with_body(r#"{"subject":"heartbeat-value","version":1,"id":4,"schema":"{\"type\":\"record\",\"name\":\"Heartbeat\",\"namespace\":\"nl.openweb.data\",\"fields\":[{\"name\":\"beat\",\"type\":\"long\"}]}"}"#)
-    ///     .create();
-    ///
-    /// let bytes = encoder.encode(vec!(("beat", Value::Long(3))), &strategy);
-    /// assert_eq!(bytes, Err(SRCError::new("Did not get a 200 response code but 404 instead", None, false).into_cache()));
-    ///
-    /// encoder.remove_errors_from_cache();
-    ///
-    /// let bytes = encoder.encode(vec!(("beat", Value::Long(3))), &strategy);
-    /// assert_eq!(bytes, Ok(vec!(0,0,0,0,4,6)))
-    /// ```
-    pub fn remove_errors_from_cache(&mut self) {
-        self.cache.retain(|_, v| v.is_ok());
-    }
-    /// Encodes a vector of values to bytes. The corrects values of the 'keys' depend on the schema
-    /// being fetched at runtime. For example you might agree on a schema with a consuming party and
-    /// /or upload a schema to the schema registry before starting the program. In the future an
-    /// 'encode with schema' might be added which makes it easier to make sure the schema will
-    /// become available in the correct way.
-    ///
-    /// ```
-    ///  # use mockito::{mock, server_address};
-    ///  # use schema_registry_converter::Encoder;
-    ///  # use schema_registry_converter::schema_registry::SubjectNameStrategy;
-    ///  # use avro_rs::types::Value;
-    ///
-    /// let _m = mock("GET", "/subjects/heartbeat-nl.openweb.data.Heartbeat/versions/latest")
-    ///     .with_status(200)
-    ///     .with_header("content-type", "application/vnd.schemaregistry.v1+json")
-    ///     .with_body(r#"{"subject":"heartbeat-value","version":1,"id":3,"schema":"{\"type\":\"record\",\"name\":\"Heartbeat\",\"namespace\":\"nl.openweb.data\",\"fields\":[{\"name\":\"beat\",\"type\":\"long\"}]}"}"#)
-    ///     .create();
-    ///
-    /// let mut encoder = Encoder::new(server_address().to_string());
-    /// let strategy = SubjectNameStrategy::TopicRecordNameStrategy("heartbeat".into(), "nl.openweb.data.Heartbeat".into());
-    /// let bytes = encoder.encode(vec!(("beat", Value::Long(3))), &strategy);
-    ///
-    /// assert_eq!(bytes, Ok(vec!(0,0,0,0,3,6)))
-    /// ```
-    pub fn encode(
-        &mut self,
-        values: Vec<(&'static str, Value)>,
-        subject_name_strategy: &SubjectNameStrategy,
-    ) -> Result<Vec<u8>, SRCError> {
-        let schema_and_id = self.get_schema_and_id(subject_name_strategy);
-        match schema_and_id {
-            Ok((schema, id)) => to_bytes(&schema, *id, values),
-            Err(e) => Err(e.clone()),
-        }
-    }
-
-    /// Encodes struct to bytes. The corrects values of the 'keys' depend on the schema being
-    /// fetched at runtime. For example you might agree on a schema with a consuming party and
-    /// /or upload a schema to the schema registry before starting the program. In the future an
-    /// 'encode with schema' might be added which makes it easier to make sure the schema will
-    /// become available in the correct way.
-    ///
-    /// ```
-    ///  # use mockito::{mock, server_address};
-    ///  # use schema_registry_converter::Encoder;
-    ///  # use schema_registry_converter::schema_registry::SubjectNameStrategy;
-    ///  # use serde::Serialize;
-    ///  # use avro_rs::types::Value;
-    ///
-    /// let _m = mock("GET", "/subjects/heartbeat-nl.openweb.data.Heartbeat/versions/latest")
-    ///     .with_status(200)
-    ///     .with_header("content-type", "application/vnd.schemaregistry.v1+json")
-    ///     .with_body(r#"{"subject":"heartbeat-value","version":1,"id":3,"schema":"{\"type\":\"record\",\"name\":\"Heartbeat\",\"namespace\":\"nl.openweb.data\",\"fields\":[{\"name\":\"beat\",\"type\":\"long\"}]}"}"#)
-    ///     .create();
-    ///
-    ///  #[derive(Serialize)]
-    ///    struct Heartbeat {
-    ///        beat: i64,
-    ///    }
-    ///
-    /// let mut encoder = Encoder::new(server_address().to_string());
-    /// let strategy = SubjectNameStrategy::TopicRecordNameStrategy("heartbeat".into(), "nl.openweb.data.Heartbeat".into());
-    /// let bytes = encoder.encode_struct(Heartbeat{beat: 3}, &strategy);
-    ///
-    /// assert_eq!(bytes, Ok(vec!(0,0,0,0,3,6)))
-    /// ```
-    pub fn encode_struct(
-        &mut self,
-        item: impl Serialize,
-        subject_name_strategy: &SubjectNameStrategy,
-    ) -> Result<Vec<u8>, SRCError> {
-        let schema_and_id = self.get_schema_and_id(subject_name_strategy);
-        match schema_and_id {
-            Ok((schema, id)) => item_to_bytes(&schema, *id, item),
-            Err(e) => Err(e.clone()),
-        }
-    }
-
-    fn get_schema_and_id(
-        &mut self,
-        subject_name_strategy: &SubjectNameStrategy,
-    ) -> &mut Result<(Schema, u32), SRCError> {
-        let schema_registry_url = &self.schema_registry_url;
-        self.cache
-            .entry(get_subject(subject_name_strategy))
-            .or_insert_with(|| {
-                match get_schema_by_subject(schema_registry_url, &subject_name_strategy) {
-                    Ok(v) => Ok(v),
-                    Err(e) => Err(e.into_cache()),
-                }
-            })
-    }
-}
-
-fn to_payload<T: ToAvro>(schema: &Schema, id: u32, record: T) -> Result<Vec<u8>, SRCError> {
-    let mut payload = vec![0u8];
-    let mut buf = [0u8; 4];
-    BigEndian::write_u32(&mut buf, id);
-    payload.extend_from_slice(&buf);
-    match to_avro_datum(schema, record) {
-        Ok(v) => {
-            payload.extend_from_slice(v.as_slice());
-            Ok(payload)
-        }
-        Err(e) => Err(SRCError::non_retryable_from_err(
-            e,
-            "Could not get avro bytes",
-        )),
-    }
-}
-
-/// Using the schema with a vector of values the values will be correctly deserialized according to
-/// the avro specification.
-fn to_bytes(
-    schema: &Schema,
-    id: u32,
-    values: Vec<(&'static str, Value)>,
-) -> Result<Vec<u8>, SRCError> {
-    let mut record = match Record::new(schema) {
-        Some(v) => v,
-        None => {
-            return Err(SRCError::new(
-                "Could not create record from schema",
-                None,
-                false,
-            ))
-        }
-    };
-    for value in values {
-        record.put(value.0, value.1)
-    }
-    to_payload(schema, id, record)
-}
-
-/// Using the schema with an item implementing serialize the item will be correctly deserialized
-/// according to the avro specification.
-fn item_to_bytes(schema: &Schema, id: u32, item: impl Serialize) -> Result<Vec<u8>, SRCError> {
-    let record = match to_value(item)
-        .map_err(|e| SRCError::non_retryable_from_err(e, "Could not transform to avro_rs value"))
-        .and_then(|r| r.fix_fixed(schema))
-        .map(|r| r.resolve(schema))
-    {
-        Ok(Ok(v)) => v,
-        Ok(Err(e)) => return Err(SRCError::non_retryable_from_err(e, "Failed to resolve")),
-        Err(e) => return Err(e),
-    };
-    to_payload(schema, id, record)
-}
-
-fn get_name(schema: &Schema) -> Name {
-    match schema {
-        Schema::Record { name: n, .. } => n.clone(),
-        _ => Name::new("no record"),
-    }
-}
-
-#[test]
-fn to_bytes_no_record() {
-    let schema = Schema::Boolean;
-    let result = to_bytes(&schema, 5, vec![("beat", Value::Long(3))]);
-    assert_eq!(
-        result,
-        Err(SRCError::new(
-            "Could not create record from schema",
-            None,
-            false,
-        ))
-    )
-}
-
-#[test]
-fn to_bytes_no_transfer_wrong() {
-    let schema = Schema::parse_str(r#"{"type":"record","name":"Name","namespace":"nl.openweb.data","fields":[{"name":"name","type":"string","avro.java.string":"String"}]}"#).unwrap();
-    let result = to_bytes(&schema, 5, vec![("beat", Value::Long(3))]);
-    assert_eq!(
-        result,
-        Err(SRCError::new(
-            "Could not get avro bytes",
-            Some("Validation error: value does not match schema"),
-            false,
-        ))
-    )
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::schema_registry::SuppliedSchema;
-    use avro_rs::from_value;
-    use mockito::{mock, server_address};
-    use serde::{Deserialize, Serialize};
-
-    #[derive(Debug, Deserialize, Serialize)]
-    struct Heartbeat {
-        beat: i64,
-    }
-
-    #[derive(Serialize)]
-    struct NoWayAvro {
-        map: HashMap<u32, u32>,
-    }
-
-    #[derive(Debug, PartialEq, Eq, Hash, PartialOrd, Ord, Clone, Deserialize, Serialize)]
-    pub enum Atype {
-        #[serde(rename = "AUTO")]
-        Auto,
-        #[serde(rename = "MANUAL")]
-        Manual,
-    }
-
-    impl Default for Atype {
-        fn default() -> Self {
-            Atype::Auto
-        }
-    }
-
-    pub type Uuid = [u8; 16];
-
-    #[serde(default)]
-    #[derive(Debug, PartialEq, Clone, Deserialize, Serialize)]
-    pub struct ConfirmAccountCreation {
-        pub id: Uuid,
-        pub a_type: Atype,
-    }
-
-    impl Default for ConfirmAccountCreation {
-        fn default() -> ConfirmAccountCreation {
-            ConfirmAccountCreation {
-                id: Uuid::default(),
-                a_type: Atype::Auto,
-            }
-        }
-    }
-
-    #[test]
-    fn display_decoder() {
-        let decoder = Decoder::new(server_address().to_string());
-        assert_eq!(
-            "Decoder { schema_registry_url: \"127.0.0.1:1234\", cache: {} }".to_owned(),
-            format!("{:?}", decoder)
-        )
-    }
-
-    #[test]
-    fn test_decoder_default() {
-        let _m = mock("GET", "/schemas/ids/1")
-            .with_status(200)
-            .with_header("content-type", "application/vnd.schemaregistry.v1+json")
-            .with_body(r#"{"schema":"{\"type\":\"record\",\"name\":\"Heartbeat\",\"namespace\":\"nl.openweb.data\",\"fields\":[{\"name\":\"beat\",\"type\":\"long\"}]}"}"#)
-            .create();
-
-        let mut decoder = Decoder::new(format!("http://{}", server_address()));
-        let heartbeat = decoder.decode(Some(&[0, 0, 0, 0, 1, 6]));
-
-        assert_eq!(
-            heartbeat,
-            Ok(Value::Record(vec![("beat".to_string(), Value::Long(3))]))
-        );
-
-        let item = match from_value::<Heartbeat>(&heartbeat.unwrap()) {
-            Ok(h) => h,
-            Err(_) => unreachable!(),
-        };
-        assert_eq!(item.beat, 3i64);
-    }
-
-    #[test]
-    fn test_decoder_with_name() {
-        let _m = mock("GET", "/schemas/ids/1")
-            .with_status(200)
-            .with_header("content-type", "application/vnd.schemaregistry.v1+json")
-            .with_body(r#"{"schema":"{\"type\":\"record\",\"name\":\"Heartbeat\",\"namespace\":\"nl.openweb.data\",\"fields\":[{\"name\":\"beat\",\"type\":\"long\"}]}"}"#)
-            .create();
-
-        let mut decoder = Decoder::new(format!("http://{}", server_address()));
-        let heartbeat = decoder.decode_with_name(Some(&[0, 0, 0, 0, 1, 6]));
-        let item = match heartbeat {
-            Ok((name, value)) => match name.name.as_str() {
-                "Heartbeat" => match name.namespace {
-                    Some(namespace) => match namespace.as_str() {
-                        "nl.openweb.data" => from_value::<Heartbeat>(&value).unwrap(),
-                        ns => panic!("Unexpected namespace {}", ns),
-                    },
-                    None => panic!("No namespace, was expected"),
-                },
-                name => panic!("Unexpected name {}", name),
-            },
-            Err(_) => unreachable!(),
-        };
-        assert_eq!(item.beat, 3i64);
-    }
-
-    #[test]
-    fn test_decoder_no_bytes() {
-        let mut decoder = Decoder::new(format!("http://{}", server_address()));
-        let heartbeat = decoder.decode(None);
-
-        assert_eq!(heartbeat, Ok(Value::Null))
-    }
-
-    #[test]
-    fn test_decoder_with_name_no_bytes() {
-        let mut decoder = Decoder::new(format!("http://{}", server_address()));
-        let heartbeat = decoder.decode_with_name(None);
-
-        assert_eq!(heartbeat, Ok((Name::new("null"), Value::Null)))
-    }
-
-    #[test]
-    fn test_decoder_magic_byte_not_present() {
-        let mut decoder = Decoder::new(format!("http://{}", server_address()));
-        let heartbeat = decoder.decode(Some(&[1, 0, 0, 0, 1, 6]));
-
-        assert_eq!(heartbeat, Ok(Value::Bytes(vec![1, 0, 0, 0, 1, 6])))
-    }
-
-    #[test]
-    fn test_decoder_with_name_magic_byte_not_present() {
-        let mut decoder = Decoder::new(format!("http://{}", server_address()));
-        let heartbeat = decoder.decode_with_name(Some(&[1, 0, 0, 0, 1, 6]));
-
-        assert_eq!(
-            heartbeat,
-            Ok((Name::new("bytes"), Value::Bytes(vec![1, 0, 0, 0, 1, 6])))
-        )
-    }
-
-    #[test]
-    fn test_decoder_not_enough_bytes() {
-        let mut decoder = Decoder::new(format!("http://{}", server_address()));
-        let heartbeat = decoder.decode(Some(&[0, 0, 0, 0]));
-
-        assert_eq!(heartbeat, Ok(Value::Bytes(vec![0, 0, 0, 0])))
-    }
-
-    #[test]
-    fn test_decoder_wrong_data() {
-        let _m = mock("GET", "/schemas/ids/1")
-            .with_status(200)
-            .with_header("content-type", "application/vnd.schemaregistry.v1+json")
-            .with_body(r#"{"schema":"{\"type\":\"record\",\"name\":\"Heartbeat\",\"namespace\":\"nl.openweb.data\",\"fields\":[{\"name\":\"beat\",\"type\":\"long\"}]}"}"#)
-            .create();
-
-        let mut decoder = Decoder::new(format!("http://{}", server_address()));
-        let heartbeat = decoder.decode(Some(&[0, 0, 0, 0, 1]));
-
-        assert_eq!(
-            heartbeat,
-            Err(SRCError::new(
-                "Could not transform bytes using schema",
-                Some("failed to fill whole buffer"),
-                false,
-            ))
-        )
-    }
-
-    #[test]
-    fn test_decoder_with_name_wrong_data() {
-        let _m = mock("GET", "/schemas/ids/1")
-            .with_status(200)
-            .with_header("content-type", "application/vnd.schemaregistry.v1+json")
-            .with_body(r#"{"schema":"{\"type\":\"record\",\"name\":\"Heartbeat\",\"namespace\":\"nl.openweb.data\",\"fields\":[{\"name\":\"beat\",\"type\":\"long\"}]}"}"#)
-            .create();
-
-        let mut decoder = Decoder::new(format!("http://{}", server_address()));
-        let heartbeat = decoder.decode_with_name(Some(&[0, 0, 0, 0, 1]));
-
-        assert_eq!(
-            heartbeat,
-            Err(SRCError::new(
-                "Could not transform bytes using schema",
-                Some("failed to fill whole buffer"),
-                false,
-            ))
-        )
-    }
-
-    #[test]
-    fn test_decoder_no_json_response() {
-        let _m = mock("GET", "/schemas/ids/1")
-            .with_status(200)
-            .with_header("content-type", "application/vnd.schemaregistry.v1+json")
-            .with_body(r#"{"type\":\"record\",\"name\":\"Heartbeat\",\"namespace\":\"nl.openweb.data\",\"fields\":[{\"name\":\"beat\",\"type\":\"long\"}]}"}"#)
-            .create();
-
-        let mut decoder = Decoder::new(format!("http://{}", server_address()));
-        let heartbeat = decoder.decode(Some(&[0, 0, 0, 0, 1, 6]));
-
-        assert_eq!(
-            heartbeat,
-            Err(SRCError::new(
-                "Invalid json string",
-                Some("expected `:` at line 1 column 130"),
-                false
-            )
-            .into_cache())
-        )
-    }
-
-    #[test]
-    fn test_decoder_with_name_no_json_response() {
-        let _m = mock("GET", "/schemas/ids/1")
-            .with_status(200)
-            .with_header("content-type", "application/vnd.schemaregistry.v1+json")
-            .with_body(r#"{"type\":\"record\",\"name\":\"Heartbeat\",\"namespace\":\"nl.openweb.data\",\"fields\":[{\"name\":\"beat\",\"type\":\"long\"}]}"}"#)
-            .create();
-
-        let mut decoder = Decoder::new(format!("http://{}", server_address()));
-        let heartbeat = decoder.decode_with_name(Some(&[0, 0, 0, 0, 1, 6]));
-
-        assert_eq!(
-            heartbeat,
-            Err(SRCError::new(
-                "Invalid json string",
-                Some("expected `:` at line 1 column 130"),
-                false
-            )
-            .into_cache())
-        )
-    }
-
-    #[test]
-    fn test_decoder_schema_registry_unavailable() {
-        let mut decoder = Decoder::new("http://bogus".to_string());
-        let heartbeat = decoder.decode(Some(&[0, 0, 0, 10, 1, 6]));
-
-        assert_eq!(
-            heartbeat,
-            Err(SRCError::new(
-                "error performing get to schema registry",
-                Some("Couldn\'t resolve host name"),
-                true,
-            )
-            .into_cache())
-        )
-    }
-
-    #[test]
-    fn test_decoder_default_no_schema_in_response() {
-        let _m = mock("GET", "/schemas/ids/1")
-            .with_status(200)
-            .with_header("content-type", "application/vnd.schemaregistry.v1+json")
-            .with_body(r#"{"no-schema":"{\"type\":\"record\",\"name\":\"Heartbeat\",\"namespace\":\"nl.openweb.data\",\"fields\":[{\"name\":\"beat\",\"type\":\"long\"}]}"}"#)
-            .create();
-
-        let mut decoder = Decoder::new(format!("http://{}", server_address()));
-        let heartbeat = decoder.decode(Some(&[0, 0, 0, 0, 1, 6]));
-
-        assert_eq!(
-            heartbeat,
-            Err(SRCError::new("Could not get raw schema from response", None, false).into_cache())
-        )
-    }
-
-    #[test]
-    fn test_decoder_default_wrong_schema_in_response() {
-        let _m = mock("GET", "/schemas/ids/1")
-            .with_status(200)
-            .with_header("content-type", "application/vnd.schemaregistry.v1+json")
-            .with_body(r#"{"schema":"{\"type\":\"record\",\"name\":\"Heartbeat\",\"namespace\":\"nl.openweb.data\"}"}"#)
-            .create();
-
-        let mut decoder = Decoder::new(format!("http://{}", server_address()));
-        let heartbeat = decoder.decode(Some(&[0, 0, 0, 0, 1, 6]));
-
-        assert_eq!(
-            heartbeat,
-            Err(SRCError::new(
-                "Could not parse schema",
-                Some("Failed to parse schema: No `fields` in record"),
-                false,
-            )
-            .into_cache())
-        )
-    }
-
-    #[test]
-    fn test_decoder_fixed_with_enum() {
-        let _m = mock("GET", "/schemas/ids/6")
-            .with_status(200)
-            .with_header("content-type", "application/vnd.schemaregistry.v1+json")
-            .with_body(r#"{"schema":"{\"type\":\"record\",\"name\":\"ConfirmAccountCreation\",\"namespace\":\"nl.openweb.data\",\"fields\":[{\"name\":\"id\",\"type\":{\"type\":\"fixed\",\"name\":\"Uuid\",\"size\":16}},{\"name\":\"a_type\",\"type\":{\"type\":\"enum\",\"name\":\"Atype\",\"symbols\":[\"AUTO\",\"MANUAL\"]}}]}"}"#)
-            .create();
-
-        let mut decoder = Decoder::new(format!("http://{}", server_address()));
-        let cac = decoder.decode(Some(&[0, 0, 0, 0, 6, 204, 240, 237, 74, 227, 188, 75, 46, 183, 163, 122, 214, 178, 72, 118, 162, 2]));
-
-        assert_eq!(cac, Ok(Value::Record(vec!(("id".to_string(), Value::Fixed(16, vec!(204, 240, 237, 74, 227, 188, 75, 46, 183, 163, 122, 214, 178, 72, 118, 162))), ("a_type".to_string(), Value::Enum(1, "MANUAL".to_string()))))));
-    }
-
-    #[test]
-    fn test_decoder_cache() {
-        let mut decoder = Decoder::new(format!("http://{}", server_address()));
-        let bytes = [0, 0, 0, 0, 2, 6];
-
-        let _m = mock("GET", "/schemas/ids/2")
-            .with_status(404)
-            .with_header("content-type", "application/vnd.schemaregistry.v1+json")
-            .with_body(r#"{"error_code":40403,"message":"Schema not found"}"#)
-            .create();
-        let heartbeat = decoder.decode(Some(&bytes));
-        assert_eq!(
-            heartbeat,
-            Err(SRCError::new(
-                "Did not get a 200 response code but 404 instead",
-                None,
-                false,
-            )
-            .into_cache())
-        );
-        let _m = mock("GET", "/schemas/ids/2")
-            .with_status(200)
-            .with_header("content-type", "application/vnd.schemaregistry.v1+json")
-            .with_body(r#"{"schema":"{\"type\":\"record\",\"name\":\"Heartbeat\",\"namespace\":\"nl.openweb.data\",\"fields\":[{\"name\":\"beat\",\"type\":\"long\"}]}"}"#)
-            .create();
-
-        let heartbeat = decoder.decode(Some(&bytes));
-        assert_eq!(
-            heartbeat,
-            Err(SRCError::new(
-                "Did not get a 200 response code but 404 instead",
-                None,
-                false,
-            )
-            .into_cache())
-        );
-
-        decoder.remove_errors_from_cache();
-
-        let heartbeat = decoder.decode(Some(&bytes));
-        assert_eq!(
-            heartbeat,
-            Ok(Value::Record(vec![("beat".to_string(), Value::Long(3))]))
-        )
-    }
-
-    #[test]
-    fn display_encode() {
-        let decoder = Encoder::new(server_address().to_string());
-        assert_eq!(
-            "Encoder { schema_registry_url: \"127.0.0.1:1234\", cache: {} }".to_owned(),
-            format!("{:?}", decoder)
-        )
-    }
-
-    #[test]
-    fn test_encode_key_and_value() {
-        let _m = mock("GET", "/subjects/heartbeat-value/versions/latest")
-            .with_status(200)
-            .with_header("content-type", "application/vnd.schemaregistry.v1+json")
-            .with_body(r#"{"subject":"heartbeat-value","version":1,"id":3,"schema":"{\"type\":\"record\",\"name\":\"Heartbeat\",\"namespace\":\"nl.openweb.data\",\"fields\":[{\"name\":\"beat\",\"type\":\"long\"}]}"}"#)
-            .create();
-
-        let _n = mock("GET", "/subjects/heartbeat-key/versions/latest")
-            .with_status(200)
-            .with_header("content-type", "application/vnd.schemaregistry.v1+json")
-            .with_body(r#"{"subject":"heartbeat-value","version":1,"id":4,"schema":"{\"type\":\"record\",\"name\":\"Name\",\"namespace\":\"nl.openweb.data\",\"fields\":[{\"name\":\"name\",\"type\":\"string\",\"avro.java.string\":\"String\"}]}"}"#)
-            .create();
-
-        let mut encoder = Encoder::new(format!("http://{}", server_address()));
-
-        let key_strategy = SubjectNameStrategy::TopicNameStrategy("heartbeat".into(), true);
-        let bytes = encoder.encode(
-            vec![("name", Value::String("Some name".to_owned()))],
-            &key_strategy,
-        );
-
-        assert_eq!(
-            bytes,
-            Ok(vec![
-                0, 0, 0, 0, 4, 18, 83, 111, 109, 101, 32, 110, 97, 109, 101,
-            ])
-        );
-
-        let value_strategy = SubjectNameStrategy::TopicNameStrategy("heartbeat".into(), false);
-        let bytes = encoder.encode(vec![("beat", Value::Long(3))], &value_strategy);
-
-        assert_eq!(bytes, Ok(vec![0, 0, 0, 0, 3, 6]))
-    }
-
-    #[test]
-    fn test_using_record_name() {
-        let _m = mock("GET", "/subjects/heartbeat-nl.openweb.data.Heartbeat/versions/latest")
-            .with_status(200)
-            .with_header("content-type", "application/vnd.schemaregistry.v1+json")
-            .with_body(r#"{"subject":"heartbeat-value","version":1,"id":3,"schema":"{\"type\":\"record\",\"name\":\"Heartbeat\",\"namespace\":\"nl.openweb.data\",\"fields\":[{\"name\":\"beat\",\"type\":\"long\"}]}"}"#)
-            .create();
-
-        let mut encoder = Encoder::new(format!("http://{}", server_address()));
-        let strategy = SubjectNameStrategy::TopicRecordNameStrategy(
-            "heartbeat".into(),
-            "nl.openweb.data.Heartbeat".into(),
-        );
-        let bytes = encoder.encode(vec![("beat", Value::Long(3))], &strategy);
-
-        assert_eq!(bytes, Ok(vec![0, 0, 0, 0, 3, 6]))
-    }
-
-    #[test]
-    fn test_encoder_no_id_in_response() {
-        let _m = mock("GET", "/subjects/heartbeat-nl.openweb.data.Heartbeat/versions/latest")
-            .with_status(200)
-            .with_header("content-type", "application/vnd.schemaregistry.v1+json")
-            .with_body(r#"{"subject":"heartbeat-value","version":1,"no-id":3,"schema":"{\"type\":\"record\",\"name\":\"Heartbeat\",\"namespace\":\"nl.openweb.data\",\"fields\":[{\"name\":\"beat\",\"type\":\"long\"}]}"}"#)
-            .create();
-
-        let mut encoder = Encoder::new(format!("http://{}", server_address()));
-        let strategy = SubjectNameStrategy::TopicRecordNameStrategy(
-            "heartbeat".into(),
-            "nl.openweb.data.Heartbeat".into(),
-        );
-        let bytes = encoder.encode(vec![("beat", Value::Long(3))], &strategy);
-
-        assert_eq!(
-            bytes,
-            Err(SRCError::new("Could not get id from response", None, false).into_cache())
-        )
-    }
-
-    #[test]
-    fn test_encoder_schema_registry_unavailable() {
-        let mut encoder = Encoder::new("http://bogus".into());
-        let strategy = SubjectNameStrategy::TopicRecordNameStrategy(
-            "heartbeat".into(),
-            "nl.openweb.data.Balance".into(),
-        );
-        let result = encoder.encode(vec![("beat", Value::Long(3))], &strategy);
-
-        assert_eq!(
-            result,
-            Err(SRCError::new(
-                "error performing get to schema registry",
-                Some("Couldn\'t resolve host name"),
-                true,
-            )
-            .into_cache())
-        )
-    }
-
-    #[test]
-    fn test_encoder_unknown_protocol() {
-        let mut encoder = Encoder::new("hxxx://bogus".into());
-        let strategy = SubjectNameStrategy::TopicRecordNameStrategy(
-            "heartbeat".into(),
-            "nl.openweb.data.Balance".into(),
-        );
-        let result = encoder.encode(vec![("beat", Value::Long(3))], &strategy);
-
-        assert_eq!(
-            result,
-            Err(SRCError::new(
-                "error performing get to schema registry",
-                Some("Unsupported protocol"),
-                true,
-            )
-            .into_cache())
-        )
-    }
-
-    #[test]
-    fn test_encoder_schema_registry_unavailable_with_record() {
-        let mut encoder = Encoder::new("http://bogus".into());
-        let heartbeat_schema = SuppliedSchema::new(r#"{"type":"record","name":"Balance","namespace":"nl.openweb.data","fields":[{"name":"beat","type":"long"}]}"#.into());
-        let strategy =
-            SubjectNameStrategy::RecordNameStrategyWithSchema(Box::from(heartbeat_schema));
-        let result = encoder.encode(vec![("beat", Value::Long(3))], &strategy);
-
-        assert_eq!(
-            result,
-            Err(SRCError::new(
-                "error performing post to schema registry",
-                Some("Couldn\'t resolve host name"),
-                true,
-            )
-            .into_cache())
-        )
-    }
-
-    #[test]
-    fn test_encode_cache() {
-        let mut encoder = Encoder::new(format!("http://{}", server_address()));
-        let strategy = SubjectNameStrategy::RecordNameStrategy("nl.openweb.data.Heartbeat".into());
-
-        let _m = mock("GET", "/subjects/nl.openweb.data.Heartbeat/versions/latest")
-            .with_status(404)
-            .with_header("content-type", "application/vnd.schemaregistry.v1+json")
-            .with_body(r#"{"error_code":40403,"message":"Schema not found"}"#)
-            .create();
-
-        let bytes = encoder.encode(vec![("beat", Value::Long(3))], &strategy);
-        assert_eq!(
-            bytes,
-            Err(SRCError::new(
-                "Did not get a 200 response code but 404 instead",
-                None,
-                false,
-            )
-            .into_cache())
-        );
-
-        let _n = mock("GET", "/subjects/nl.openweb.data.Heartbeat/versions/latest")
-            .with_status(200)
-            .with_header("content-type", "application/vnd.schemaregistry.v1+json")
-            .with_body(r#"{"subject":"heartbeat-value","version":1,"id":4,"schema":"{\"type\":\"record\",\"name\":\"Heartbeat\",\"namespace\":\"nl.openweb.data\",\"fields\":[{\"name\":\"beat\",\"type\":\"long\"}]}"}"#)
-            .create();
-
-        let bytes = encoder.encode(vec![("beat", Value::Long(3))], &strategy);
-        assert_eq!(
-            bytes,
-            Err(SRCError::new(
-                "Did not get a 200 response code but 404 instead",
-                None,
-                false,
-            )
-            .into_cache())
-        );
-
-        encoder.remove_errors_from_cache();
-
-        let bytes = encoder.encode(vec![("beat", Value::Long(3))], &strategy);
-        assert_eq!(bytes, Ok(vec![0, 0, 0, 0, 4, 6]))
-    }
-
-    #[test]
-    fn test_encode_key_and_value_supplied_record() {
-        let _n = mock("POST", "/subjects/heartbeat-key/versions")
-            .with_status(200)
-            .with_header("content-type", "application/vnd.schemaregistry.v1+json")
-            .with_body(r#"{"id":3}"#)
-            .create();
-
-        let _m = mock("POST", "/subjects/heartbeat-value/versions")
-            .with_status(200)
-            .with_header("content-type", "application/vnd.schemaregistry.v1+json")
-            .with_body(r#"{"id":4}"#)
-            .create();
-
-        let mut encoder = Encoder::new(format!("http://{}", server_address()));
-
-        let name_schema = SuppliedSchema::new(r#"{"type":"record","name":"Name","namespace":"nl.openweb.data","fields":[{"name":"name","type":"string","avro.java.string":"String"}]}"#.into());
-        let key_strategy = SubjectNameStrategy::TopicNameStrategyWithSchema(
-            "heartbeat".into(),
-            true,
-            Box::from(name_schema),
-        );
-        let bytes = encoder.encode(
-            vec![("name", Value::String("Some name".to_owned()))],
-            &key_strategy,
-        );
-        assert_eq!(
-            bytes,
-            Ok(vec![
-                0, 0, 0, 0, 3, 18, 83, 111, 109, 101, 32, 110, 97, 109, 101,
-            ])
-        );
-        let heartbeat_schema = SuppliedSchema::new(r#"{"type":"record","name":"Heartbeat","namespace":"nl.openweb.data","fields":[{"name":"beat","type":"long"}]}"#.into());
-        let value_strategy = SubjectNameStrategy::TopicNameStrategyWithSchema(
-            "heartbeat".into(),
-            false,
-            Box::from(heartbeat_schema),
-        );
-        let bytes = encoder.encode(vec![("beat", Value::Long(3))], &value_strategy);
-        assert_eq!(bytes, Ok(vec![0, 0, 0, 0, 4, 6]))
-    }
-
-    #[test]
-    fn test_encode_record_name_strategy_supplied_record() {
-        let _n = mock("POST", "/subjects/nl.openweb.data.Heartbeat/versions")
-            .with_status(200)
-            .with_header("content-type", "application/vnd.schemaregistry.v1+json")
-            .with_body(r#"{"id":11}"#)
-            .create();
-
-        let mut encoder = Encoder::new(format!("http://{}", server_address()));
-
-        let heartbeat_schema = SuppliedSchema::new(r#"{"type":"record","name":"Heartbeat","namespace":"nl.openweb.data","fields":[{"name":"beat","type":"long"}]}"#.into());
-        let strategy =
-            SubjectNameStrategy::RecordNameStrategyWithSchema(Box::from(heartbeat_schema));
-        let bytes = encoder.encode(vec![("beat", Value::Long(3))], &strategy);
-        assert_eq!(bytes, Ok(vec![0, 0, 0, 0, 11, 6]))
-    }
-
-    #[test]
-    fn test_encode_record_name_strategy_supplied_record_wrong_response() {
-        let _n = mock("POST", "/subjects/nl.openweb.data.Heartbeat/versions")
-            .with_status(200)
-            .with_header("content-type", "application/vnd.schemaregistry.v1+json")
-            .with_body(r#"{"no-id":11}"#)
-            .create();
-
-        let mut encoder = Encoder::new(format!("http://{}", server_address()));
-
-        let heartbeat_schema = SuppliedSchema::new(r#"{"type":"record","name":"Heartbeat","namespace":"nl.openweb.data","fields":[{"name":"beat","type":"long"}]}"#.into());
-        let strategy =
-            SubjectNameStrategy::RecordNameStrategyWithSchema(Box::from(heartbeat_schema));
-        let bytes = encoder.encode(vec![("beat", Value::Long(3))], &strategy);
-        assert_eq!(
-            bytes,
-            Err(SRCError::new("Could not get id from response", None, false).into_cache())
-        )
-    }
-
-    #[test]
-    fn test_encode_topic_record_name_strategy_supplied_record() {
-        let _n = mock("POST", "/subjects/hb-nl.openweb.data.Heartbeat/versions")
-            .with_status(200)
-            .with_header("content-type", "application/vnd.schemaregistry.v1+json")
-            .with_body(r#"{"id":23}"#)
-            .create();
-
-        let mut encoder = Encoder::new(format!("http://{}", server_address()));
-
-        let heartbeat_schema = SuppliedSchema::new(r#"{"type":"record","name":"Heartbeat","namespace":"nl.openweb.data","fields":[{"name":"beat","type":"long"}]}"#.into());
-        let strategy = SubjectNameStrategy::TopicRecordNameStrategyWithSchema(
-            "hb".into(),
-            Box::from(heartbeat_schema),
-        );
-        let bytes = encoder.encode(vec![("beat", Value::Long(3))], &strategy);
-        assert_eq!(bytes, Ok(vec![0, 0, 0, 0, 23, 6]))
-    }
-
-    #[test]
-    fn test_encode_topic_record_name_strategy_schema_registry_not_available() {
-        let mut encoder = Encoder::new(format!("http://{}", server_address()));
-
-        let heartbeat_schema = SuppliedSchema::new(r#"{"type":"record","name":"Heartbeat","namespace":"nl.openweb.data","fields":[{"name":"beat","type":"long"}]}"#.into());
-        let strategy = SubjectNameStrategy::TopicRecordNameStrategyWithSchema(
-            "hb".into(),
-            Box::from(heartbeat_schema),
-        );
-        let error = encoder.encode(vec![("beat", Value::Long(3))], &strategy);
-        assert_eq!(
-            error,
-            Err(SRCError::new(
-                "Did not get a 200 response code but 501 instead",
-                None,
-                false,
-            )
-            .into_cache())
-        )
-    }
-
-    #[test]
-    fn item_to_bytes_no_tranfer_wrong() {
-        let schema = Schema::parse_str(r#"{"type":"record","name":"Name","namespace":"nl.openweb.data","fields":[{"name":"name","type":"string","avro.java.string":"String"}]}"#).unwrap();
-
-        let result = crate::item_to_bytes(&schema, 5, Heartbeat { beat: 3 });
-        assert_eq!(
-            result,
-            Err(SRCError::new(
-                "error fixing record",
-                Some("missing field name in record"),
-                false,
-            ))
-        )
-    }
-
-    #[test]
-    fn derive_but_not_valid_avro() {
-        let schema = Schema::parse_str(r#"{"type":"record","name":"Name","namespace":"nl.openweb.data","fields":[{"name":"name","type":"string","avro.java.string":"String"}]}"#).unwrap();
-        let mut map = HashMap::new();
-        map.insert(1, 2);
-        let result = crate::item_to_bytes(&schema, 5, NoWayAvro { map });
-        assert_eq!(
-            result,
-            Err(SRCError::new(
-                "Could not transform to avro_rs value",
-                Some("map key is not a string"),
-                false,
-            ))
-        )
-    }
-
-    #[test]
-    fn item_to_bytes_proper_bytes_record_with_fixed() {
-        let schema = Schema::parse_str(r#"{"type":"record","name":"ConfirmAccountCreation","namespace":"nl.openweb.data","fields":[{"name":"id","type":{"type":"fixed","name":"Uuid","size":16}},{"name":"a_type","type":{"type":"enum","name":"Atype","symbols":["AUTO","MANUAL"]}}]}"#).unwrap();
-        let item = ConfirmAccountCreation {
-            id: [
-                204, 240, 237, 74, 227, 188, 75, 46, 183, 163, 122, 214, 178, 72, 118, 162,
-            ],
-            a_type: Atype::Manual,
-        };
-        let result = crate::item_to_bytes(&schema, 6, item);
-        assert_eq!(
-            result,
-            Ok(vec!(0, 0, 0, 0, 6, 204, 240, 237, 74, 227, 188, 75, 46, 183, 163, 122, 214, 178, 72, 118, 162, 2))
-        )
+    pub fn avro(schema_registry_url: String) -> AvroEncoder {
+        AvroEncoder::new(schema_registry_url)
     }
 }
