+++ tests/integration_tests.rs
@@ -3,11 +3,12 @@ extern crate schema_registry_converter;
 mod kafka_consumer;
 mod kafka_producer;
 
-use crate::kafka_consumer::{consume, DeserializedRecord};
+use crate::kafka_consumer::{consume_avro, DeserializedAvroRecord};
 use crate::kafka_producer::get_producer;
 use avro_rs::types::Value;
 use rand::Rng;
-use schema_registry_converter::schema_registry::{SubjectNameStrategy, SuppliedSchema};
+use rdkafka::message::FromBytes;
+use schema_registry_converter::schema_registry::{SchemaType, SubjectNameStrategy, SuppliedSchema};
 
 fn get_schema_registry_url() -> String {
     "http://localhost:8081".into()
@@ -18,11 +19,18 @@ fn get_brokers() -> &'static str {
 }
 
 fn get_heartbeat_schema() -> Box<SuppliedSchema> {
-    Box::from(SuppliedSchema::new(r#"{"type":"record","name":"Heartbeat","namespace":"nl.openweb.data","fields":[{"name":"beat","type":"long"}]}"#.into()))
+    Box::from(SuppliedSchema {
+        name: String::from("nl.openweb.data.Heartbeat"),
+        schema_type: SchemaType::AVRO,
+        schema: String::from(
+            r#"{"type":"record","name":"Heartbeat","namespace":"nl.openweb.data","fields":[{"name":"beat","type":"long"}]}"#,
+        ),
+        references: vec![],
+    })
 }
 
-fn test_beat_value(key_value: i64, value_value: i64) -> Box<dyn Fn(DeserializedRecord) -> ()> {
-    Box::new(move |rec: DeserializedRecord| {
+fn test_beat_value(key_value: i64, value_value: i64) -> Box<dyn Fn(DeserializedAvroRecord) -> ()> {
+    Box::new(move |rec: DeserializedAvroRecord| {
         println!("testing record {:#?}", rec);
         let key_values = match rec.key {
             Value::Record(v) => v,
@@ -45,21 +53,25 @@ fn test_beat_value(key_value: i64, value_value: i64) -> Box<dyn Fn(DeserializedR
     })
 }
 
-fn do_test(topic: &str, key_strategy: SubjectNameStrategy, value_strategy: SubjectNameStrategy) {
+fn do_avro_test(
+    topic: &str,
+    key_strategy: SubjectNameStrategy,
+    value_strategy: SubjectNameStrategy,
+) {
     let mut rng = rand::thread_rng();
     let key_value = rng.gen::<i64>();
     let value_value = rng.gen::<i64>();
     let mut producer = get_producer(get_brokers(), get_schema_registry_url());
     let key_values = vec![("beat", Value::Long(key_value))];
     let value_values = vec![("beat", Value::Long(value_value))];
-    producer.send(
+    producer.send_avro(
         topic,
         key_values,
         value_values,
         key_strategy,
         value_strategy,
     );
-    consume(
+    consume_avro(
         get_brokers(),
         "test",
         get_schema_registry_url(),
@@ -82,7 +94,7 @@ fn test1_topic_name_strategy_with_schema() {
         false,
         get_heartbeat_schema(),
     );
-    do_test(topic, key_strategy, value_strategy)
+    do_avro_test(topic, key_strategy, value_strategy)
 }
 
 #[test]
@@ -91,7 +103,7 @@ fn test2_record_name_strategy_with_schema() {
     let topic = "recordnamestrategy";
     let key_strategy = SubjectNameStrategy::RecordNameStrategyWithSchema(get_heartbeat_schema());
     let value_strategy = SubjectNameStrategy::RecordNameStrategyWithSchema(get_heartbeat_schema());
-    do_test(topic, key_strategy, value_strategy)
+    do_avro_test(topic, key_strategy, value_strategy)
 }
 
 #[test]
@@ -106,7 +118,7 @@ fn test3_topic_record_name_strategy_with_schema() {
         topic.into(),
         get_heartbeat_schema(),
     );
-    do_test(topic, key_strategy, value_strategy)
+    do_avro_test(topic, key_strategy, value_strategy)
 }
 
 #[test]
@@ -115,7 +127,7 @@ fn test4_topic_name_strategy_schema_now_available() {
     let topic = "topicnamestrategy";
     let key_strategy = SubjectNameStrategy::TopicNameStrategy(topic.into(), true);
     let value_strategy = SubjectNameStrategy::TopicNameStrategy(topic.into(), false);
-    do_test(topic, key_strategy, value_strategy)
+    do_avro_test(topic, key_strategy, value_strategy)
 }
 
 #[test]
@@ -125,7 +137,7 @@ fn test5_record_name_strategy_schema_now_available() {
     let key_strategy = SubjectNameStrategy::RecordNameStrategy("nl.openweb.data.Heartbeat".into());
     let value_strategy =
         SubjectNameStrategy::RecordNameStrategy("nl.openweb.data.Heartbeat".into());
-    do_test(topic, key_strategy, value_strategy)
+    do_avro_test(topic, key_strategy, value_strategy)
 }
 
 #[test]
@@ -140,5 +152,76 @@ fn test6_topic_record_name_strategy_schema_now_available() {
         topic.into(),
         "nl.openweb.data.Heartbeat".into(),
     );
-    do_test(topic, key_strategy, value_strategy)
+    do_avro_test(topic, key_strategy, value_strategy)
+}
+
+#[test]
+#[cfg_attr(not(feature = "kafka_test"), ignore)]
+fn test7_test_avro_from_java_test_app() {
+    let topic = "testavro";
+    let test = Box::new(move |rec: DeserializedAvroRecord| {
+        println!("testing record {:#?}", rec);
+        let key_bytes = match rec.key {
+            Value::Bytes(v) => v,
+            _ => panic!("Keys wasn't bytes"),
+        };
+        let key_string = match str::from_bytes(key_bytes.as_ref()) {
+            Ok(v) => v,
+            _ => panic!("Could not create string key successfully"),
+        };
+        assert_eq!("testkey", key_string, "check string key");
+        let value_values = match rec.value {
+            Value::Record(v) => v,
+            _ => panic!("Not a record, while only only those expected"),
+        };
+        let id_key = match &value_values[0] {
+            (_id, Value::Fixed(16, _v)) => _id,
+            _ => panic!("Not a fixed value of 16 bytes while that was expected"),
+        };
+        assert_eq!("id", id_key, "expected id key to be id");
+        let enum_value = match &value_values[1] {
+            (_id, Value::Enum(0, v)) => v,
+            _ => panic!("Not an enum value for by while that was expected"),
+        };
+        assert_eq!("Java", enum_value, "expect message from Java");
+        let counter_value = match &value_values[2] {
+            (_id, Value::Long(v)) => v,
+            _ => panic!("Not a long value for counter while that was expected"),
+        };
+        assert_eq!(&1i64, counter_value, "counter is 1");
+        let input_value = match &value_values[3] {
+            (_id, Value::Union(v)) => v,
+            _ => panic!("Not an unions value for input while that was expected"),
+        };
+        assert_eq!(
+            &Box::new(Value::String(String::from("String"))),
+            input_value,
+            "Optional string is string"
+        );
+        let results = match &value_values[4] {
+            (_id, Value::Array(v)) => v,
+            _ => panic!("Not an array value for results while that was expected"),
+        };
+        let result = match results.get(0).expect("one item to be present") {
+            Value::Record(v) => v,
+            _ => panic!("Not record for first of results while that was expected"),
+        };
+        let up_result = match &result[0] {
+            (_id, Value::String(v)) => v,
+            _ => panic!("First result value wasn't a string"),
+        };
+        assert_eq!("STRING", up_result, "expected upper case string");
+        let down_result = match &result[1] {
+            (_id, Value::String(v)) => v,
+            _ => panic!("Second result value wasn't a string"),
+        };
+        assert_eq!("string", down_result, "expected upper case string");
+    });
+    consume_avro(
+        get_brokers(),
+        "test",
+        get_schema_registry_url(),
+        &vec![topic],
+        test,
+    )
 }
