+++ tests/kafka_producer.rs
@@ -1,16 +1,17 @@
 use avro_rs::types::Value;
 use rdkafka::config::ClientConfig;
 use rdkafka::producer::{FutureProducer, FutureRecord};
+use schema_registry_converter::avro::AvroEncoder;
 use schema_registry_converter::schema_registry::SubjectNameStrategy;
 use schema_registry_converter::Encoder;
 
 pub struct RecordProducer {
     producer: FutureProducer,
-    encoder: Encoder,
+    avro_encoder: AvroEncoder,
 }
 
 impl<'a> RecordProducer {
-    pub fn send(
+    pub fn send_avro(
         &'a mut self,
         topic: &'a str,
         key_values: Vec<(&'static str, Value)>,
@@ -18,11 +19,11 @@ impl<'a> RecordProducer {
         key_strategy: SubjectNameStrategy,
         value_strategy: SubjectNameStrategy,
     ) {
-        let payload = match self.encoder.encode(value_values, &value_strategy) {
+        let payload = match self.avro_encoder.encode(value_values, &value_strategy) {
             Ok(v) => v,
             Err(e) => panic!("Error getting payload: {}", e),
         };
-        let key = match self.encoder.encode(key_values, &key_strategy) {
+        let key = match self.avro_encoder.encode(key_values, &key_strategy) {
             Ok(v) => v,
             Err(e) => panic!("Error getting payload: {}", e),
         };
@@ -46,6 +47,9 @@ pub fn get_producer(brokers: &str, schema_registry_url: String) -> RecordProduce
         .set("queue.buffering.max.messages", "10")
         .create()
         .expect("Producer creation error");
-    let encoder = Encoder::new(schema_registry_url);
-    RecordProducer { producer, encoder }
+    let avro_encoder = Encoder::new(schema_registry_url);
+    RecordProducer {
+        producer,
+        avro_encoder,
+    }
 }
